Fixed hanging, Added Scrolling
The hanging was a result of windows being confused that the program was not receiving any input. A new error has now appeared where with certain levels of zoom, the cells become vertically missaligned by half a cell. I suspect a rounding issue or maybe an off by one error perhaps. Also implemented an exit key, and removed math.sqrt as a dependency.